packages/nft-res-generator/generator/pyproject.toml
[tool.poetry]
name = "celts"
version = "0.1.0"
description = ""
authors = ["Mint Road Ltd."]

[tool.poetry.dependencies]
python = "^3.8.10"
pandas = "^1.4.0"
Pillow = "^9.0.0"
selenium = "^4.1.0"
webdriver-manager = "^3.5.2"
python-dotenv = "^0.19.2"
PyVirtualDisplay = "^3.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
